audit: security hardening, cost optimization, and best practices
Security:
- K8s: add securityContext (runAsNonRoot, drop ALL caps, readOnlyRootFilesystem)
- K8s: add NetworkPolicy (default deny + allow frontend→backend→postgres)
- K8s: encrypt EBS volumes (StorageClass encrypted=true)
- Frontend: run nginx as non-root user (port 8080, pid in /tmp)
- Backend: add alembic.ini + migrations to Docker image (K8s init container fix)
- Backend: Float → Numeric(10,2) for money fields (no floating-point rounding)
- Nginx: fix security headers lost in nested location blocks

Best practices:
- Frontend Dockerfile: node:18 → node:20 (LTS)
- EKS module: default k8s version 1.31 → 1.32, fix stale variable descriptions
- K8s frontend: NLB annotation scheme=internet-facing for public subnet placement
- K8s backend: fix stale GHCR comment → ECR
- deploy.sh: add network policy step (6 steps total)
- docker-compose.prod.yml: frontend port 80:8080 (matches non-root nginx)

+++ backend/app/models/game.py
@@ -1,5 +1,5 @@
 import uuid
-from sqlalchemy import Boolean, Column, String, Float, DateTime, ForeignKey, Integer
+from sqlalchemy import Boolean, Column, String, Numeric, DateTime, ForeignKey, Integer
 from sqlalchemy import Uuid
 from sqlalchemy.orm import relationship
 from datetime import datetime, timezone
@@ -17,7 +17,7 @@ class Game(Base):
     id = Column(Uuid(as_uuid=True), primary_key=True, default=uuid.uuid4)
     user_id = Column(Uuid(as_uuid=True), ForeignKey("users.id"), nullable=False)
     status = Column(String, default="active", nullable=False)  # active, finished
-    bet_amount = Column(Float, nullable=False)
+    bet_amount = Column(Numeric(10, 2), nullable=False)
     result = Column(
         String, nullable=True
     )  # win, lose, push, blackjack (comma-sep for split)
